UP-10 Added pedantic flag and fixed warnings

+++ include/glyphs/column.h
@@ -22,7 +22,7 @@ public:
 
     void Insert(ColumnPtr& column);
     void InsertBack(std::list<GlyphPtr>&& glyphs);
-    void Insert(size_t insertPosition, std::list<GlyphPtr>&& itemsToInsert);
+    void Insert(int insertPosition, std::list<GlyphPtr>&& itemsToInsert);
 
     bool IsEmpty() const;
     bool IsFull() const;
